Fixed error in how LastDiv was being stored as int instead of decimal

+++ api/Dtos/Stock/FMPStock.cs
@@ -12,7 +12,7 @@ namespace api.Dtos
         public double beta { get; set; }
         public int volAvg { get; set; }
         public long mktCap { get; set; }
-        public int lastDiv { get; set; }
+        public decimal lastDiv { get; set; }
         public string range { get; set; }
         public double changes { get; set; }
         public string companyName { get; set; }
